refactor canvas/editor
