--- ui/src/features/Canvas/components/index.ts
@@ -1,8 +1,0 @@
-export * from "./CustomConnectionLine";
-export * from "./Nodes";
-export * from "./Toolbox";
-export * from "./Infobar";
-export * from "./CanvasActionbar";
-export * from "./Actionbar";
-export * from "./Breadcrumb";
-export * from "./CanvasTabs";
