+++ ui/src/features/Editor/components/Nodes/GeneralNode/types.ts
