+++ ui/src/features/Editor/components/Nodes/NoteNode/index.tsx
