+++ ui/src/features/Editor/components/Nodes/WriterNode.tsx
