+++ ui/src/features/Editor/components/OverlayUI/components/index.ts
@@ -1,0 +1,6 @@
+export * from "./ActionBar";
+export * from "./Breadcrumb";
+export * from "./CanvasTabs";
+export * from "./CanvasActionbar";
+export * from "./Infobar";
+export * from "./Toolbox";
