+++ ui/src/routes/workspace_.$workspaceId_.project.$projectId.lazy.tsx
@@ -1,6 +1,6 @@
 import { createLazyFileRoute } from "@tanstack/react-router";
 
-import Canvas from "@flow/features/Canvas";
+import Canvas from "@flow/features/Editor";
 import { useCurrentProject } from "@flow/stores";
 
 export const Route = createLazyFileRoute("/workspace/$workspaceId/project/$projectId")({
